server/pyproject.toml
[tool.poetry]
name = "bus-me"
version = "0.1.0"
description = "Bus Me backend"
authors = ["Shane Duan <s_duan@u.pacific.edu>"]

[tool.poetry.dependencies]
python = "^3.6"
eventlet = "^0.24.1"
pony = "^0.7.9"
config2 = "^0.3.1"
aiohttp = "^3.5"
python-socketio = "^4.0"
pyjwt = "^1.7"
cryptography = "^2.6"
websockets = "^7.0"
aiocache = "^0.10.1"
attributedict = "^0.3.0"

[tool.poetry.dev-dependencies]
black = {version = "^18.3-alpha.0",allows-prereleases = true}
pylint = "^2.3"

[tool.poetry.scripts]
start = "bus_me:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
